requirements.txt
chromadb==1.1.0
sentence-transformers==5.1.1
openai==1.108.1
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
lxml==4.9.3
python-dotenv==1.0.0